pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goodshyt-governance"
version = "0.1.0"
description = "Ethical validation, policy enforcement, guardrails, and transparency auditing for accountable intelligent systems."
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["fastapi>=0.115.0", "pydantic>=2.8.0", "pyyaml>=6.0.1", "uvicorn>=0.30.0"]
[project.optional-dependencies]
dev = ["pytest>=8.3.0", "mypy>=1.10.0", "ruff>=0.6.0", "httpx>=0.27.0"]
[tool.setuptools.packages.find]
where = ["src"]
[tool.ruff]
line-length = 100
[tool.mypy]
python_version = "3.11"
strict = true
